# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Survey Panda API",
        "description": "The Survey Panda API developed by Frank Li.",
        "version": "0.2.0",
        "contact": {
            "email": "threestones2@hotmail.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3001"
        }
    ],
    "basePath": "/",
...
    "paths": {
...
        "/company/new": {
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "Create new company record with default questionset",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Cheeky Bikes"
                                    },
                                    "industry": {
                                        "type": "string",
                                        "example": "Retail"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/company/update": {
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "Update existing company's questionset",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "example": "cheekybikes"
                                    },
                                    "measures": {
                                        "type": ["string"],
                                        "example": ["Workload Perception", "Stress Level"]
                                    },
                                    "questions": {
                                        "type": ["string"],
                                        "example": ["How was the workload today?", "Did you find today a bit stressful?"]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/company/find": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Find a company's information base on nameid",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "example": "examplecompany"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}